[image-builder-mk3]: expose config
